Fix install instructions
This project uses `main` instead of `master` and also has no release tags, so mint tries to install from a non-existing master branch and fails. Specifying `@main` seems to fix the install and well as the mint run command. As an alternative, you could make a release tag, but I thought making a PR would be more helpful for discussion.